requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
beautifulsoup4==4.12.2
requests==2.31.0
textblob==0.17.1
torch==2.1.2
transformers==4.35.2
numpy==1.26.4
pandas==2.1.3
gtts==2.4.0
python-dotenv==1.0.0
httpx==0.21.0
httpcore==0.14.0
h11==0.12.0
deep-translator==1.11.4